+++ Changelog.md
@@ -3,6 +3,7 @@
 ## NEXT
 
 - Fixes #635: Skills and Attributes which appear in `@Buff` enrichers are now localized when the default label is created.
+- Ensure Effects Panel is updated when the effects on the selected actor change.
 
 ## 13.15.0